EXCEL_SEARCH_NON_TABULAR/requirements.txt
openpyxl==3.1.5
pandas==2.3.2
Office365-REST-Python-Client==2.6.2
psycopg2-binary==2.9.10
python-dotenv==1.1.1
watchdog==6.0.0
numpy==2.3.3
pywin32==306; platform_system == "Windows"